requirements.txt
absl-py==1.0.0
astor==0.8.1
astroid==2.10.0
azure-core==1.23.0        
azure-functions==1.9.0    
azure-storage-blob==12.6.0
black==22.1.0
cached-property==1.5.2    
certifi==2021.10.8        
cffi==1.15.0
chardet==3.0.4
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
decorator==5.1.1
flake8==4.0.1
gast==0.2.2
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==2.10
importlib-metadata==4.2.0
isodate==0.6.1
isort==5.10.1
Keras==2.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.3.1
Markdown==3.1.1
mccabe==0.6.1
msrest==0.6.21
mypy-extensions==0.4.3
networkx==2.3
nltk==3.4
numpy==1.21.5
oauthlib==3.2.0
opencv-python==3.4.9.33
opt-einsum==3.3.0
pathspec==0.9.0
Pillow==6.0.0
platformdirs==2.5.1
protobuf==3.9.2
py==1.7.0
pycodestyle==2.4.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.3.0
pylint==2.1.1
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.24.0
requests-oauthlib==1.3.1
scipy==1.7.3
singledispatch==3.4.0.3
six==1.11.0
tensorboard==1.15.0
tensorflow==1.15.3
tensorflow-estimator==1.15.1
termcolor==1.1.0
tomli==2.0.1
typed-ast==1.5.2
typing_extensions==4.1.1
urllib3==1.25.11
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0